src/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
openai>=1.0.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0
